fix: 알림페이지 렌더링 로직
