+++ apps/web/src/app/(mobile)/alarm/hooks/useNotificationsViewModel.ts
@@ -7,10 +7,11 @@ import { NotificationQueries } from '@/entities/notification/notification.querie
 const limit = 20;
 
 export const useNotificationsViewModel = () => {
-  const { data, isLoading, error, hasNextPage, fetchNextPage, isFetchingNextPage } =
-    useInfiniteQuery(NotificationQueries.infiniteNotifications({ limit }));
+  const { data, isLoading, hasNextPage, fetchNextPage, isFetchingNextPage } = useInfiniteQuery(
+    NotificationQueries.infiniteNotifications({ limit }),
+  );
 
-  const notifications = useMemo(() => (data?.pages ?? []).flat(), [data?.pages]);
+  const notifications = useMemo(() => data?.pages.flatMap((page) => page) ?? [], [data]);
 
   const noData = !isLoading && notifications.length === 0;
 
@@ -22,8 +23,6 @@ export const useNotificationsViewModel = () => {
     },
   });
 
-  const isNotLogin = (error as any)?.response?.errors?.[0]?.extensions?.code === 'FORBIDDEN';
-
   useEffect(() => {
     // noop: kept for parity and potential side-effects later
   }, [notifications]);
@@ -31,7 +30,6 @@ export const useNotificationsViewModel = () => {
   return {
     notifications,
     loading: isLoading || isFetchingNextPage,
-    isNotLogin,
     noData,
     hasNextData: !!hasNextPage,
     ref,
