+++ apps/web/src/app/(mobile)/alarm/page.tsx
@@ -1,12 +1,12 @@
 import BasicLayout from '@/components/layout/BasicLayout';
 import { NAV_TYPE } from '@/components/layout/BottomNav';
 
-import AlarmList from './components/AlarmList';
+import AlarmContainer from './components/AlarmContainer';
 
 const Alarm = () => {
   return (
     <BasicLayout title="알림" hasBottomNav navType={NAV_TYPE.ALARM}>
-      <AlarmList />
+      <AlarmContainer />
     </BasicLayout>
   );
 };
