+++ apps/web/src/shared/api/notification/notification.service.ts
@@ -13,6 +13,13 @@ export class NotificationService {
   }
 
   static async getNotifications(variables: QueryNotificationsQueryVariables) {
-    return execute(QueryNotificationsDocument, variables).then((res) => res.data);
+    return execute(QueryNotificationsDocument, variables).then((res) => {
+      // @ts-expect-error errors type
+      if (res.errors) {
+        // @ts-expect-error errors type
+        throw res.errors;
+      }
+      return res.data.notifications;
+    });
   }
 }
