+++ apps/web/src/shared/hooks/useIsLoggedIn.ts
@@ -1,11 +1,14 @@
 'use client';
 
+import { atom, useAtom } from 'jotai';
 import { useEffect, useState } from 'react';
 
 import { getAccessToken } from '@/app/actions/token';
 
+const accessTokenAtom = atom<string | null>(null);
+
 const useIsLoggedIn = () => {
-  const [accessToken, setAccessToken] = useState<string>();
+  const [accessToken, setAccessToken] = useAtom(accessTokenAtom);
   const [isLoading, setIsLoading] = useState(true);
 
   useEffect(() => {
@@ -13,7 +16,7 @@ const useIsLoggedIn = () => {
       setIsLoading(true);
       try {
         const token = await getAccessToken();
-        setAccessToken(token);
+        setAccessToken(token ?? null);
       } catch (error) {
         console.error('Failed to fetch access token:', error);
       } finally {
@@ -21,9 +24,9 @@ const useIsLoggedIn = () => {
       }
     };
     fetchAccessToken();
-  }, []);
+  }, [setAccessToken]);
 
-  return !isLoading && !!accessToken;
+  return { isLoggedIn: !isLoading && !!accessToken, isLoading };
 };
 
 export default useIsLoggedIn;
